pyproject.toml
[project]
name = "mcp-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "mcp[cli]>=1.8.0",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.61",
]
